Initial commit: add AWS Terraform provider and S3 modules
This commit introduces the AWS provider for Terraform in a new 'providers.tf' file. Additionally, two modules are added to the 'main.tf' which create S3 buckets for an email blast webapp and lambda. A '.gitignore' file is also included to prevent Terraform state and plan files from being committed.