Update root record in Route53
Modified the aws_route53_record resource in the 'route53_zone_records_github.tf' file. The name has been changed from "www" to an empty string to represent the root or apex record, rather than the www subdomain. This is an essential setup for DNS settings configuration, aiding in correctly routing the traffic.