Created IAM group and users for Github Actions in AWS
This commit adds a new IAM group "GitHubActions" for AWS, as well as related IAM policies for access to S3, Cloudfront, and Lambda services. Also, created a user "GithubActions" and added the user to the aforementioned group. These changes are to provide Github Actions with necessary permissions to interact with AWS resources.