Add AWS ACM certificate and update Cloudfront viewer certificate
The commit introduces an ACM certificate for the domain in a new file 'acm_certificate_dynamdev.tf'. It also updates the Cloudfront distribution in 'cloudfront_distribution_dynamdev.tf' to use the newly added certificate instead of a default one. Furthermore, a Route53 DNS record is added for domain validation, supporting SSL/TLS for Cloudfront distribution. Additionally, AWS provider configuration is updated to specify the region.