Refactor Terraform files for email-blast-composer
This commit refactors multiple Terraform files related to the email-blast-composer module. It does so by removing aws_s3_bucket and its policy, introducing variables for route management and ACM certificate, and renaming and updating cloudfront_distribution and route53_zone records. Hence, it optimizes the configuration for better management and flexibility.