Refactor domain name routing in Terraform
This commit refactors the modules.tf and output.tf files in the Terraform production and global directories, respectively. Changes include extracting the routing domain name as a variable from the remote state output instead of hardcoding it, and adding an output for the domain name in the global output.tf file. This enhances configuration flexibility and maintainability.