Add script to list S3 zip contents and update Terraform files
A new bash script has been added to list the contents within an S3 zip file. Multiple Terraform files have also been updated or added to integrate the usage of this script. This includes updates to the AWS Lambda setup and IAM roles, so Lambdas can use the new script and correctly handle the data it provides.