Add script to get zip file hash from S3 in app-email-blast-composer
Added a new script `get_s3_zip_code_hash.sh` to calculate the ETag (MD5 hash) of an uploaded zip file in an S3 bucket. Updated the Terraform configurations to integrate this script and further optimized the sorting of AWS Lambda functions. This implementation enhances the handling of zip files within the app-email-blast-composer module.