Refactor subdomain variables in Terraform modules
Renamed the variable for the subdomain name in multiple Terraform modules and added a new variable for the api subdomain. This reconfigures the AWS route53 records and updates the alias in AWS Cloudfront to match the new variable name. This change better distinguishes between app routes and api routes, enhancing readability and maintainability of the code.