Update Lambda functions and resources in AWS modules
This commit updates the naming structure of several AWS Lambda functions and resources for improvement in the AWS modules. It adds new resources for the 'queue_email' and improves the arrangement of permissions in the 'lambda.tf' file. Also, it includes slight modifications in variables 'api_gateway_domain' and 'lambda_functions' in 'local.tf' file. The changes do not affect existing functionality but enhance overall module operation and readability.