Add new AWS SQS queue and corresponding route in module
This commit introduces a new AWS SQS queue named 'email' in the 'app-email-blast-composer' module. To make it accessible, the 'route_queue_sub_domain_name' variable was added and updated in the production configuration. Also, it includes permissions for 'sqs:SendMessage' in the IAM role policies, enabling function interaction with the new queue. Some unused lambda permissions in 'lambda.tf' are removed as well.