Refactor AWS CloudWatch logs and update IAM policies
This revision introduces essential changes to the AWS cloudwatch logs by renaming the existing log group and adding log groups for 'queue_email' and 'send_email' functions. The IAM policies are also updated with new permissions for the aforementioned log groups, enhancing their operational capabilities. An AWS caller identity data source is added in the 'data.tf' file and the calling identity's Account ID is updated as the output value in the 'output.tf'.