Add README.md for Personal Terraform Setup
Added a README.md outlining the steps to get started with the using the personal Terraform setup. The README includes a tutorial on cloning the repository, initializing the Terraform environment, reviewing the plan, and applying the configuration. It also informs users that the setup is for personal use and may not suit everyone's needs.